feat: "방 API 구현"
이미지 삭제 전 유저 검증하도록 수정
Rooms delete 기능 추가
Rooms가 "delete" 되도 실제 삭제된게 아님으로 CascadeType.REMOVE, orphanRemoval = true 제거

Related to: #6

+++ src/main/java/com/example/airdns/domain/room/service/RoomsServiceImplV1.java
@@ -1,10 +1,15 @@
 package com.example.airdns.domain.room.service;
 
 import com.example.airdns.domain.equipment.service.EquipmentsService;
+import com.example.airdns.domain.image.entity.Images;
 import com.example.airdns.domain.image.service.ImagesService;
 import com.example.airdns.domain.room.converter.RoomsConverter;
-import com.example.airdns.domain.room.dto.RoomsRequestDto.*;
-import com.example.airdns.domain.room.dto.RoomsResponseDto.*;
+import com.example.airdns.domain.room.dto.RoomsRequestDto.CreateRoomsRequestDto;
+import com.example.airdns.domain.room.dto.RoomsRequestDto.ReadRoomsListRequestDto;
+import com.example.airdns.domain.room.dto.RoomsRequestDto.UpdateRoomsImagesRequestDto;
+import com.example.airdns.domain.room.dto.RoomsRequestDto.UpdateRoomsRequestDto;
+import com.example.airdns.domain.room.dto.RoomsResponseDto.ReadRoomsResponseDto;
+import com.example.airdns.domain.room.dto.RoomsResponseDto.UpdateRoomsImagesResponseDto;
 import com.example.airdns.domain.room.entity.Rooms;
 import com.example.airdns.domain.room.exception.RoomsCustomException;
 import com.example.airdns.domain.room.exception.RoomsExceptionCode;
@@ -12,7 +17,6 @@ import com.example.airdns.domain.room.repository.RoomsRepository;
 import com.example.airdns.domain.roomequipment.service.RoomEquipmentsService;
 import com.example.airdns.domain.user.entity.Users;
 import com.example.airdns.domain.user.enums.UserRole;
-import com.example.airdns.global.awss3.S3FileUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -44,7 +48,6 @@ public class RoomsServiceImplV1 implements RoomsService {
         updateEquipments(rooms, requestDto.getEquipment());
         uploadImages(rooms, files);
 
-
         return RoomsConverter.toDto(rooms);
     }
 
@@ -72,6 +75,7 @@ public class RoomsServiceImplV1 implements RoomsService {
         return RoomsConverter.toDto(requestDto, rooms);
     }
 
+    @Transactional
     @Override
     public UpdateRoomsImagesResponseDto updateRoomsImages(
             UpdateRoomsImagesRequestDto requestDto,
@@ -80,14 +84,15 @@ public class RoomsServiceImplV1 implements RoomsService {
             Users users) {
         Rooms rooms = findById(roomsId);
 
-        //TODO 데이터 없으면 리턴
+        validateUserIsRoomsHost(rooms, users);
 
-        if (!rooms.getUsers().getId().equals(users.getId())) {
-            throw new RoomsCustomException(RoomsExceptionCode.NO_PERMISSION_USER);
+        if (requestDto != null) {
+            deleteImage(rooms, requestDto.getRemoveImages());
         }
 
-        deleteImage(rooms, requestDto.getRemoveImages());
-        uploadImages(rooms, files);
+        if (files != null) {
+            uploadImages(rooms, files);
+        }
 
         return RoomsConverter.toImagesDto(rooms);
 
@@ -95,7 +100,11 @@ public class RoomsServiceImplV1 implements RoomsService {
 
     @Override
     public void deleteRooms(Long roomsId, Users users) {
+        Rooms rooms = findById(roomsId);
 
+        validateUserIsRoomsHost(rooms, users);
+
+        roomsRepository.delete(rooms);
     }
 
     @Override
@@ -114,6 +123,12 @@ public class RoomsServiceImplV1 implements RoomsService {
                 .orElseThrow(() -> new RoomsCustomException(RoomsExceptionCode.INVALID_ROOMS_ID));
     }
 
+    private void validateUserIsRoomsHost(Rooms rooms, Users users) {
+        if (!rooms.getUsers().getId().equals(users.getId())) {
+            throw new RoomsCustomException(RoomsExceptionCode.NO_PERMISSION_USER);
+        }
+    }
+
     private void updateEquipments(Rooms rooms, List<Long> equipments) {
         equipments.stream().distinct().forEach(
                 equipment -> rooms.addEquipments(
@@ -126,12 +141,19 @@ public class RoomsServiceImplV1 implements RoomsService {
 
     private void uploadImages(Rooms rooms, List<MultipartFile> files) {
         for(MultipartFile file : files) {
+            if (file == null) continue;
+
             //TODO 롤백 시 이미지 제거 (선택1: 롤백 로직 추가, 선택2: 배치 시스템 구성)
             rooms.addImage(imagesService.createImages(rooms, file));
         }
     }
 
     private void deleteImage(Rooms rooms, List<Long> removeImages) {
+        if (!rooms.getImagesList().stream().map(Images::getId).toList()
+                .containsAll(removeImages)) {
+            throw new RoomsCustomException(RoomsExceptionCode.IMAGES_NOT_EXIST);
+        }
+
         removeImages.forEach(images -> imagesService.deleteImages(images, rooms));
     }
 }
